Move client.Batch to its own file.
Replaced client.batcher with client.B() which plays nicer with godoc and
removes any confusion as to whether the Batch created is specific to a
DB or Txn.

Fixes #1254.
