+++ acceptance/single_key_test.go
@@ -95,7 +95,7 @@ func TestSingleKey(t *testing.T) {
 					if err := v.UnmarshalBinary(r.ValueBytes()); err != nil {
 						return err
 					}
-					return txn.Commit(txn.B.Put(key, v+1))
+					return txn.Commit(client.B().Put(key, v+1))
 				})
 				if err != nil {
 					resultCh <- result{err: err}
