+++ client/db_test.go
@@ -140,7 +140,7 @@ func ExampleBatch() {
 	s, db := setup()
 	defer s.Stop()
 
-	b := db.B.Get("aa").Put("bb", "2")
+	b := client.B().Get("aa").Put("bb", "2")
 	if err := db.Run(b); err != nil {
 		panic(err)
 	}
@@ -159,7 +159,7 @@ func ExampleDB_Scan() {
 	s, db := setup()
 	defer s.Stop()
 
-	b := db.B.Put("aa", "1").Put("ab", "2").Put("bb", "3")
+	b := client.B().Put("aa", "1").Put("ab", "2").Put("bb", "3")
 	if err := db.Run(b); err != nil {
 		panic(err)
 	}
@@ -180,7 +180,7 @@ func ExampleDB_Del() {
 	s, db := setup()
 	defer s.Stop()
 
-	if err := db.Run(db.B.Put("aa", "1").Put("ab", "2").Put("ac", "3")); err != nil {
+	if err := db.Run(client.B().Put("aa", "1").Put("ab", "2").Put("ac", "3")); err != nil {
 		panic(err)
 	}
 	if err := db.Del("ab"); err != nil {
@@ -204,13 +204,13 @@ func ExampleTx_Commit() {
 	defer s.Stop()
 
 	err := db.Txn(func(txn *client.Txn) error {
-		return txn.Commit(txn.B.Put("aa", "1").Put("ab", "2"))
+		return txn.Commit(client.B().Put("aa", "1").Put("ab", "2"))
 	})
 	if err != nil {
 		panic(err)
 	}
 
-	b := db.B.Get("aa").Get("ab")
+	b := client.B().Get("aa").Get("ab")
 	if err := db.Run(b); err != nil {
 		panic(err)
 	}
@@ -292,10 +292,9 @@ func TestDebugName(t *testing.T) {
 
 func TestCommonMethods(t *testing.T) {
 	batchType := reflect.TypeOf(&client.Batch{})
-	batcherType := reflect.TypeOf(client.DB{}.B)
 	dbType := reflect.TypeOf(&client.DB{})
 	txnType := reflect.TypeOf(&client.Txn{})
-	types := []reflect.Type{batchType, batcherType, dbType, txnType}
+	types := []reflect.Type{batchType, dbType, txnType}
 
 	type key struct {
 		typ    reflect.Type
