+++ server/status/feed_test.go
@@ -231,7 +231,7 @@ func TestServerNodeEventFeed(t *testing.T) {
 
 	// Add some data in a transaction
 	err = db.Txn(func(txn *client.Txn) error {
-		return txn.Commit(txn.B.Put("a", "asdf").Put("c", "jkl;"))
+		return txn.Commit(client.B().Put("a", "asdf").Put("c", "jkl;"))
 	})
 	if err != nil {
 		t.Fatalf("error putting data to db: %s", err)
